\name{Chunk-class}
\docType{class}
\alias{Chunk-class}
\title{Chunk Class}
\description{Chunk class}
\section{Objects from the Class}{
  \code{new('Chunk',
    id = [numeric],
    dependency = [numeric],
    morphem.list = [list])}
}

\section{Slots}{
	 \describe{
    \item{\code{id}:}{chunk id}
    \item{\code{dependency}:}{dependency id}
    \item{\code{morphem.list}:}{Morphem object list}
  }
}
\section{Methods}{}

\author{Yohei Sato}
\seealso{
  \code{\link{DAService}}
}
\keyword{classes}