\name{initYjdnJlp}
\alias{initYjdnJlp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
initalize YjdnJlp Object
}
\description{
initalize YjdnJlp Object
}
\usage{
initYjdnJlp(appid)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{appid}{
Application ID for Yahoo! Japan Developer Network.
}
}
\details{

}
\value{
YjdnJlp object
}
\references{
https://e.developer.yahoo.co.jp
}
\author{
Yohei Sato
}
\note{

}



\seealso{

}
\examples{
   \dontrun{
       con <- initYjdnJlp()
   }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
