% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlp_parse.R
\name{jlp_parse}
\alias{jlp_parse}
\title{Use Parse API}
\usage{
jlp_parse(
  sentence,
  results = "ma",
  filter = c(1:13),
  uniq_by_baseform = TRUE,
  app_id = Sys.getenv("YJDN_CLIENT_ID")
)
}
\arguments{
\item{sentence}{character vector.}

\item{results}{character scalar. choose one of \code{ma} or \code{uniq}.}

\item{filter}{numeric vector.}

\item{uniq_by_baseform}{logical.}

\item{app_id}{character scalar.}
}
\value{
return results as a list.
}
\description{
Use Parse API
}
\seealso{
\url{https://developer.yahoo.co.jp/webapi/jlp/ma/v1/parse.html}
}
