\name{Keyphrase-class}
\docType{class}
\alias{Keyphrase-class}
\title{Keyphrase Class}
\description{Keyphrase class}
\section{Objects from the Class}{
  \code{new('Keyphrase',
    keyphrase = [character],
    score = [numeric])}
}

\section{Slots}{
	 \describe{
    \item{\code{keyphrase}:}{keyphrase}
    \item{\code{score}:}{score}
  }
}

\section{Methods}{}
\author{Yohei Sato}
\seealso{
  \code{\link{Keyphrase}}
}
\keyword{classes}