\name{Morphem-class}
\docType{class}
\alias{Morphem-class}
\title{Morphem Class}
\description{Morphem class}
\section{Objects from the Class}{
  \code{new('Morphem',
    surface = [character],
    reading = [character],
    baseform = [character],
    pos = [character],
    feature = [list])}
}

\section{Slots}{
	 \describe{
    \item{\code{surface}:}{surface}
    \item{\code{reading}:}{reading}
    \item{\code{baseform}:}{baseform}
    \item{\code{pos}:}{POS}
    \item{\code{feature}:}{feature list}
  }
}
\section{Methods}{}
\author{Yohei Sato}
\seealso{
  \code{\link{DAService}}
}
\keyword{classes}