\name{Word-class}
\docType{class}
\alias{Word-class}
\title{Word Class}
\description{Word class}
\section{Objects from the Class}{
  \code{new('Word',
    surface = [character],
    reading = [character],
    pos = [character],
    baseform = [character],
    count = [numeric])}
}

\section{Slots}{
	 \describe{
    \item{\code{surface}:}{surface}
    \item{\code{reading}:}{reading}
    \item{\code{pos}:}{POS}
    \item{\code{baseform}:}{baseform}
    \item{\code{count}:}{count}
  }
}
\section{Methods}{}

\author{Yohei Sato}
\seealso{
  \code{\link{MAService}}
}
\keyword{classes}