% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yjdnjlp-package.R
\docType{package}
\name{yjdnjlp}
\alias{yjdnjlp}
\alias{yjdnjlp-package}
\title{yjdnjlp: An R Client of Yahoo!Japan Text Analysis API}
\description{
Rennovation of cran/YjdnJlp.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/yjdnjlp}
  \item Report bugs at \url{https://github.com/paithiov909/yjdnjlp/issues}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
